Created app_genre and app_music_album
