+++ data/genre.json
@@ -6,5 +6,13 @@
   {
     "id" : 9743,
     "name" : "hiphop"
+  },
+  {
+    "id" : 5566,
+    "name" : "soul"
+  },
+  {
+    "id" : 3544,
+    "name" : "soul"
   }
 ]
